ion: Add compatibility with legacy Ion v1 ABI
Kernel 4.12 revamped the Ion ABI [1], rendering it incompatible with blobs
made to work with the old "v1" ABI. This commit implements the old ABI by
mapping legacy Ion handles to dma-buf fds and adding simple implementations
of the ioctls that no longer exist. The alloc ioctl's arguments struct has
changed and none of the other ioctls conflict, so this implementation is
compatible with blobs using both the v1 and v2 ABIs.

dma-buf allocations are automatically freed when no references to the fd
remain, so the FREE ioctl doesn't need to do anything meaningful. However,
newer blobs using libion will call FREE with a handle of 0 to check which
ABI the kernel implements and decide whether to use v1 or v2. Because we
want to use v2 as much as possible for efficiency, FREE needs to return
-ENOTTY to pretend that it doesn't exist if it detects that the call was
made by libion for probing the ABI, so it is not truly a no-op.

Support for cache maintenance operations and 32-bit clients will be added
later.

[1] https://source.android.com/devices/architecture/kernel/ion_abi_changes

Co-authored-by: dianlujitao <dianlujitao@lineageos.org>
Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Subhajeet Muhuri <subhajeet.muhuri@gmail.com>
Signed-off-by: sohamxda7 <sensoham135@gmail.com>
Signed-off-by: ImPrashantt <prashant33968@gmail.com>